Moved DurationMetricsCollector and sample_duration to its own module.
